refine(onboarding): tighten the welcome screens after device review
Adjustments from the first device pass.

The step strip and the trust badges now share one card treatment at a
matching s12 inset — both wrap grouped, structured content, while
prose stays on the canvas. s12 rather than s16: the strip splits its
width across five nodes and the longest label ellipsises if squeezed
harder. Carding the badges means re-inking them back to inkBody and
goldDeep, since paper copy on a paper card is invisible; that part of
the previous commit is deliberately undone. The badge card takes the
warm goldPending ground the reassurance pill already uses rather than
plain white — closer to the brand. The strip card stays paper: its
gold12 nodes would flatten against a gold ground.

The reassurance pill centres its contents rather than shrink-wrapping.
The copy runs long enough in English to fill the width anyway, so a
fitted pill would leave the two languages visibly different shapes.

Several groups sat too low with their lines packed. Each panel is
bottom-anchored under an Expanded hero, so the lever is the gap ABOVE
the nav row, not the padding below it: widening that gap grows the
panel upward and lifts the copy while the controls keep their distance
from the screen edge. Padding underneath would have carried the
controls up with it. Mediation also spreads its heading, body and pill
apart; essence lifts its portrait meta clear of the photo edge.

Next stops being a bare chevron and takes its word, via the existing
common.next. It is pinned to 104dp so the control keeps one footprint
in both languages: a user's memory of where the button sits outlives a
few pixels of tighter fit. That variable width breaks the old centring
trick, which reserved a fixed slot for a hidden control, so the row
becomes three slots with equal flex either side — the dots now sit
centred whatever the label does. Back stays a circle; it is the
secondary move.

The nav row test needed an EasyLocalization host once next resolved a
label, and that package re-exports intl's TextDirection over the
dart:ui one the layout assertions need — hence the hide.

Scope held: QeranButton gains a call site, not a change, and no next
affordance outside onboarding is touched.

+++ lib/features/onboarding/presentation/widgets/frames/onboarding_blurred_profile_card.dart
@@ -29,10 +29,12 @@ class OnboardingBlurredProfileCard extends StatelessWidget {
           OnboardingCardTopScrim(),
           OnboardingCardBottomScrim(),
           _LockBadge(),
+          // Sits clear of the photo's lower edge so the name and chips read as
+          // part of the portrait rather than crowding whatever follows it.
           Positioned(
             left: QeranSpacing.s16,
             right: QeranSpacing.s16,
-            bottom: QeranSpacing.s16,
+            bottom: QeranSpacing.s32,
             child: OnboardingProfileMetaOverlay(),
           ),
         ],
